+++ packages/patrol_cli/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
 - Download only Chromium instead of all default Playwright browsers during web runner setup. (#3156)
-- Add `--record-video` flag to `patrol test` and `patrol develop` to record a video per test case on Android emulators and iOS simulators. (#2741)
+- Add `--record-video` flag to `patrol test` and `patrol develop` to record a video per test case on Android emulators and iOS simulators. May also work on physical Android devices, depending on the vendor (some OEM ROMs block `screenrecord`); iOS physical devices are not supported. (#2741)
 
 ## 4.5.1